+++ .gitignore
@@ -1,3 +1,2 @@
-libs/
 config/
 TODO
