+++ libs/scm/ui.lua
@@ -1,0 +1,9 @@
+---@class SCMUI
+local UI = {}
+
+do
+    function UI:listScripts()
+    end
+end
+
+return UI
